+++ DarkMoon/src/DarkMoon/Camera/PerspectiveCameraController.cpp
@@ -91,7 +91,7 @@ namespace DarkMoon
 
 	bool PerspectiveCameraController::OnWindowResized(WindowResizeEvent& e)
 	{
-		OnResize(e.GetWidth(), e.GetHeight());
+		OnResize(static_cast<float>(e.GetWidth()), static_cast<float>(e.GetHeight()));
 		return false;
 	}
 }
